Fix code style

+++ modules/task_1/maximova_i_dijkstra_algorithm/dijkstra_algorithm.cpp
@@ -1,12 +1,13 @@
 // Copyright 2020 Maximova Irina
 #include "../../../modules/task_1/maximova_i_dijkstra_algorithm/dijkstra_algorithm.h"
+#include <limits.h>
 #include <omp.h>
 #include <ctime>
 #include <random>
 #include <stdexcept>
 #include <vector>
 #include <utility>
-#include <limits.h>
+
 
 Graph::Graph(int _numVertex, int _numEdges) {
   if (_numVertex <= 0)
@@ -19,7 +20,7 @@ Graph::Graph(int _numVertex, int _numEdges) {
     throw std::runtime_error("Too many edges in the graph");
   numEdges = _numEdges;
 
-  std::vector<int> vecEx(numVertex, INT_MAX);
+  std::vector<int> vecEx(numVertex, INT8_MAX);
   linkedList.resize(numVertex);
   for (int i = 0; i < numVertex; ++i) linkedList[i] = vecEx;
 }
@@ -66,7 +67,7 @@ std::vector<std::vector<int>> Graph::getLinkedList() const {
 
 int minDistVertex(const std::vector<int>& dist, const std::vector<bool>& mark) {
   int minDistV;
-  int minDist = INT_MAX;
+  int minDist = INT8_MAX;
   for (size_t i = 0; i < dist.size(); ++i)
     if (mark[i] == false && dist[i] < minDist) {
       minDist = dist[i];
@@ -79,7 +80,7 @@ std::vector<int> SeqDijkstraAlg(const Graph& graph, int sourceVertex) {
   std::vector<std::vector<int>> linkedList = graph.getLinkedList();
   int numVertex = linkedList.size();
 
-  std::vector<int> dist(numVertex, INT_MAX);
+  std::vector<int> dist(numVertex, INT8_MAX);
   std::vector<bool> mark(numVertex, false);
   int vertex;
   dist[sourceVertex] = 0;
@@ -88,7 +89,7 @@ std::vector<int> SeqDijkstraAlg(const Graph& graph, int sourceVertex) {
     vertex = minDistVertex(dist, mark);
     mark[vertex] = 1;
     for (int j = 0; j < numVertex; ++j)
-      if (!mark[j] && linkedList[vertex][j] != INT_MAX &&
+      if (!mark[j] && linkedList[vertex][j] != INT8_MAX &&
           (dist[vertex] + linkedList[vertex][j] < dist[j]))
         dist[j] = dist[vertex] + linkedList[vertex][j];
   }
